Finalized UI Structure * Filtered unneccessary objects,empty objects and duplicates from ticket form parameters
