+++ src/views/Tickets/styles/ticketOptionStyle.jsx
